Compute camera info from a main script (+)
Here, I implelemented the functions to access to MP3D-FPE dataset, iterate for each of the rooms in a scene,
call the respective function to estimate camera height, and save that information in a geoemtry file.
The definition of the room is given by the list of rooms_scenes at data/mp3d_fpe/*.json files.