Fix  mvl-dir as input
In the prev. version, the usser should explicitly pass the geometry_info or img dierectory to create a scene_list.
Here, the user will need to pass the mvl-directory (Internally the code will list geoemtry_info). This makes more sense since
the user wants a scene_list for the mvl-data and this referes to both geoemtry_info  and imgs directories.